refactor(consts): ワールドピクセル座標を consts.Pixel に統一し変換ヘルパーを追加する
同じ「ワールドのピクセル位置」が Camera では Coord[float64]、vision では Coord[consts.Pixel] と
不統一だった。Camera.Pos/Target と vision の raycast を Coord[consts.Pixel] に揃える。

タイル→ワールドピクセル中心の変換が spawn/location/camera/vision で重複していたので
consts.TileCenterToWorld に集約する。スクリーン座標のための consts.ScreenPixel 型と、
ワールド→スクリーン変換 consts.WorldToScreen も追加する（次コミットで HUD が使う）。

値は不変で golden も変わらない。serde は Pixel=float64 なので JSON 互換。全テストと lint で確認。

+++ internal/systems/render_sprite.go
@@ -68,7 +68,7 @@ func setTranslate(world w.World, op *ebiten.DrawImageOptions, camera *gc.Camera)
 
 	// カメラ位置の設定
 	if camera != nil {
-		op.GeoM.Translate(-camera.Pos.X, -camera.Pos.Y)
+		op.GeoM.Translate(-float64(camera.Pos.X), -float64(camera.Pos.Y))
 		op.GeoM.Scale(camera.Scale, camera.Scale)
 	}
 	// 画面の中央
@@ -81,7 +81,7 @@ func viewportTileBounds(world w.World, margin consts.Tile, camera *gc.Camera) (m
 	var cameraX, cameraY float64
 	cameraScale := 1.0
 	if camera != nil {
-		cameraX, cameraY, cameraScale = camera.Pos.X, camera.Pos.Y, camera.Scale
+		cameraX, cameraY, cameraScale = float64(camera.Pos.X), float64(camera.Pos.Y), camera.Scale
 	}
 	if cameraScale <= 0 {
 		cameraScale = 1.0
@@ -456,8 +456,8 @@ func (sys *RenderSpriteSystem) renderDarkness(world w.World, screen *ebiten.Imag
 	var cameraX, cameraY float64
 	cameraScale := 1.0
 	if camera != nil {
-		cameraX = camera.Pos.X
-		cameraY = camera.Pos.Y
+		cameraX = float64(camera.Pos.X)
+		cameraY = float64(camera.Pos.Y)
 		cameraScale = camera.Scale
 	}
